requirements.txt
# Django and REST Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0

# Task Queue and Caching
celery>=5.3.0
redis>=4.5.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# Environment and Configuration
python-decouple>=3.8
python-dotenv>=1.0.0

# API Documentation
drf-spectacular>=0.26.0

# Development and Testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0
coverage>=7.2.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0

# Utilities
Pillow>=10.0.0
requests>=2.31.0